feat: 1:初始化

+++ bj-business/src/main/java/com/ch/business/member/type/domain/MemberType.java
@@ -30,7 +30,7 @@ public class MemberType implements Serializable {
     /**
      * 会员折扣
      */
-    private BigDecimal memberDiscount;
+    private Long memberDiscount;
     /**
      * 备注
      */
